重命名 ERNIE-4.0-8K 模型示例代码文件并更新列表
