+++ list.md
@@ -4,7 +4,7 @@
 | 2  | 8k-preview.py       | ERNIE-3.5-8K-preview模型示例代码       | [MB/ERNIE35/8k-preview.py](MB/ERNIE35/8k-preview.py)             |
 | 3  | 128k.py             | ERNIE-3.5-128K模型示例代码             | [MB/ERNIE35/128K.py](MB/ERNIE35/128K.py)                         |
 | 4  | 128k-preview.py     | ERNIE-3.5-128K-preview模型示例代码     | [MB/ERNIE35/128k_preview.py](MB/ERNIE35/128k_preview.py)         |
-| 5  | 8k.py               | ERNIE-4.0-8K模型示例代码               | [MB/ERNIE40/8k.py](MB/ERNIE40/8k.py)                             |
+| 5  | ernie40_8k.py       | ERNIE-4.0-8K模型示例代码               | [MB/ERNIE40/ernie40_8k.py](MB/ERNIE40/ernie40_8k.py)             |
 | 6  | 8k-latest.py        | ERNIE-4.0-8K-latest模型示例代码        | [MB/ERNIE40/8k-latest.py](MB/ERNIE40/8k-latest.py)               |
 | 7  | 8k-preview.py       | ERNIE-4.0-8K-preview模型示例代码       | [MB/ERNIE40/8k-preview.py](MB/ERNIE40/8k-preview.py)             |
 | 8  | turbo-8k.py         | ERNIE-4.0-Turbo-8K模型示例代码         | [MB/ERNIE40/turbo-8k.py](MB/ERNIE40/turbo-8k.py)                 |
